[DEVELOPMENT ENVIRONMENT] * NEW: ImageEditor: Tons of bug fixes. The current selection and the   current pasted image can be moved, rotated and shrinked with the mouse. * NEW: Image rotation tool.
[GB.DRAW]
* BUG: The PaintMatrix class is now really autonomous. It can be used 
  without a current drawing device.
* NEW: PaintMatrix.Copy() is a new method that copies a matrix.
* NEW: PaintMatrix.Map() is a new method that applies a matrix to a point
  and returns the result.

[GB.FORM]
* NEW: ImageView can take the focus.
* BUG: DirChooser should not crash anymore when it cannot access a 
  bookmark.

[GB.GTK]
* NEW: Implementation of the new PaintMatrix design.

[GB.IMAGE]
* BUG: Image.Fuzzy() argument cannot be greater than 254.

[GB.QT4]
* NEW: Implementation of the new PaintMatrix design.


git-svn-id: svn://localhost/gambas/trunk@5779 867c0c6c-44f3-4631-809d-bfa615b0a4ec

+++ main/lib/geom/gb.geom.h
@@ -70,7 +70,7 @@ typedef
 	struct {
 		int version;
 		GEOM_POINT *(*CreatePoint)(int x, int y);
-		GEOM_POINTF *(*CreatePointF)(int x, int y);
+		GEOM_POINTF *(*CreatePointF)(double x, double y);
 		GEOM_RECT *(*CreateRect)(void);
 		GEOM_RECTF *(*CreateRectF)(void);
 		}
